feat: create requirement assessment read-only page
