+++ frontend/src/routes/(app)/compliance-assessments/[id=uuid]/TreeViewItemContent.svelte
@@ -105,32 +105,32 @@
 <div class="flex flex-row justify-between space-x-8">
 	<div class="flex flex-1 justify-center max-w-[80ch] flex-col">
 		<span style="font-weight: 300;">
-			{#if assessable && canEditRequirementAssessment}
+			{#if assessable}
 				<span class="w-full h-full flex rounded-token hover:text-primary-500">
-					<a href="/requirement-assessments/{ra_id}?next={$page.url.pathname}">
-						{#if title}
-							<span style="font-weight: 600;">{title}</span>
-						{/if}
-						{#if description}
-							<p>{description}</p>
-						{/if}
-					</a>
+					{#if canEditRequirementAssessment}
+						<a href="/requirement-assessments/{ra_id}/edit?next={$page.url.pathname}">
+							{#if title}
+								<span style="font-weight: 600;">{title}</span>
+							{/if}
+							{#if description}
+								<p>{description}</p>
+							{/if}
+						</a>
+					{:else}
+						<a href="/requirement-assessments/{ra_id}?next={$page.url.pathname}">
+							{#if title}
+								<span style="font-weight: 600;">{title}</span>
+							{/if}
+							{#if description}
+								<p>{description}</p>
+							{/if}
+						</a>
+					{/if}
 				</span>
 			{:else}
 				<p class="max-w-[80ch] whitespace-pre-line">
 					{#if title}
 						<span style="font-weight: 600;">{title}</span>
-						{#each Object.entries(complianceStatusColorMap) as status}
-							{#if resultCounts[status[0]]}
-								<span
-									class="badge mr-1"
-									style="background-color: {status[1] + '44'}; color: {darkenColor(status[1], 0.3)}"
-								>
-									{resultCounts[status[0]]}
-									{m[toCamelCase(status[0])]()}
-								</span>
-							{/if}
-						{/each}
 					{/if}
 					{#if description}
 						<p>{description}</p>
@@ -215,6 +215,17 @@
 	</div>
 	{#if hasAssessableChildren}
 		<div class="flex max-w-96 grow items-center space-x-2">
+			{#each Object.entries(complianceStatusColorMap) as status}
+				{#if resultCounts[status[0]]}
+					<span
+						class="badge mr-1"
+						style="background-color: {status[1] + '44'}; color: {darkenColor(status[1], 0.3)}"
+					>
+						{resultCounts[status[0]]}
+						{m[toCamelCase(status[0])]()}
+					</span>
+				{/if}
+			{/each}
 			<div
 				class="flex max-w-96 grow bg-gray-200 rounded-full overflow-hidden h-4 shrink self-center"
 			>
